refactor: extract shared statement-execution loop from the spark executor
